ENH: Moved I/O methods for meshes and surfaces and extract_mesh to contour_tools.
